PF submission function; Updated weighted scores

+++ pull-assignments.py
@@ -4,7 +4,7 @@ from openpyxl.styles import NamedStyle, Font, Alignment
 
 from lxml.html import fromstring
 from analysis  import get_weighted_scores
-from utils     import login, secrets
+from utils     import login
 
 import datetime
 import os
@@ -21,9 +21,6 @@ ASSIGNMENT_LINK_XPATH = '//h2/a[contains(.,"Download submitted assignment")]'
 ASSIGNMENT_LINKS_XPATH = '//a[contains(@class, "taskButton")]'
 RUBRIC_TABLE_XPATH = '//table[contains(@class, "rubricView")]'
 STUDENT_NAME_XPATH = '//div[@class="checkbox"]/label/div/a'
-
-USERNAME = secrets()['username'] if secrets() else ''
-PASSWORD = secrets()['password'] if secrets() else ''
 
 
 def populate_spreadsheet(assignment_name, assignments={}):
@@ -89,7 +86,8 @@ def populate_spreadsheet(assignment_name, assignments={}):
 
     print('Saving grades to %s' % workbook_path)
     if os.path.exists(workbook_path):
-        question = input('Do you wish to overwrite %s? (y/n)' % workbook_path)
+        question = input("Do you wish to overwrite %s?  (y/n)" % (workbook_path))
+
         if question == 'y':
             wb.save(workbook_path)
         else:
@@ -106,7 +104,7 @@ def pull_assignments(session):
 
     assignment_name = tree.xpath(ASSIGNMENT_NAME_XPATH)[0].text.title()
 
-    print('Pulling assignments for %s as %s...' % (assignment_name, USERNAME))
+    print('Pulling assignments for %s...' % (assignment_name))
 
     links = tree.xpath(ASSIGNMENT_LINKS_XPATH)
 
@@ -147,9 +145,10 @@ def pull_assignments(session):
         if not os.path.exists(filepath):
             os.makedirs(filepath)
         filename = filepath + st_name.replace(' ', '') + '.pdf'
-        resp = session.get(dl_link)
-        with open(filename, 'wb') as f:
-            f.write(resp.content)
+        if not os.path.exists(filename):
+            resp = session.get(dl_link)
+            with open(filename, 'wb') as f:
+                f.write(resp.content)
 
     with open('assignments/%s/Data/assignments.json' % assignment_name, 'w') as file:
         json.dump(tasks, file)
@@ -168,7 +167,7 @@ def pull_assignments(session):
 
 
 def process():
-    session = login(USERNAME, PASSWORD)
+    session = login()
     pull_assignments(session)
 
 if __name__ == "__main__":
